slight reformatting
Former-commit-id: 21deca89d04829b8d0924b7758616d2e48ba7659
Former-commit-id: c39bc65ee7faa4cc52652a765f0bef23f8c82cc0

+++ backend/src/pytorch/InterpolateArchs/GMFSS/GMFSS.py
@@ -21,7 +21,7 @@ class GMFSS(nn.Module):
         self,
         model_path,
         model_type: str = "union",
-        scale: int = 1,
+        scale: float = 1.,
         ensemble: bool = False,
         width: int = 1920,
         height: int = 1080,
